Identify tracks primarly using their URL instead of fingerprint
This has the downside that moving a file on the filesystem will likely cause an import error.
It has the upside that the subset of files which end up with different fingerprints on my laptop vs raspberry pi (despite running in a container) won't cause import errors.
Once the whole library has been scanned on the raspberry pi, this change could be reverted, because hopefully from then on fingerprints with be consistent.  (Though I've thought that fingerprints would be consistent across various previous changes and that's proved not to be the case, so we'll see).